move datastore auth tests to pytest
